% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.pred.R
\docType{class}
\name{class.pred-class}
\alias{class.pred-class}
\title{S4 class for the estimation results in the mixed SDE with random effects in the drift, in the diffusion or both}
\description{
S4 class for the estimation results in the mixed SDE with random effects in the drift, in the diffusion or both
}
\section{Slots}{

\describe{
\item{\code{estim}}{object of class Fit.class containing the results of the model estimation}

\item{\code{phipred}}{matrix of simulated values for the random effects in the drift that are used for prediction (dimensions)}

\item{\code{Xpred}}{matrix of simulated trajectories used for prediction (dimensions)}

\item{\code{idexpred}}{vector of indexes of the true trajectories that are used for prediction}
}}

