% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodGamma.R
\name{contrastGamma}
\alias{contrastGamma}
\title{Contrast based on the Euler approximation of the likelihood for parameter estimation when there is one random 
effect in the diffusion coefficient and no random effect in the drift.}
\usage{
contrastGamma(a, lambda, U, V, S, K, drift.fixed)
}
\arguments{
\item{a}{value of the shape of the Gamma distribution.}

\item{lambda}{value of the scape of the Gamma distribution.}

\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{drift.fixed}{values of the fixed effects in the drift.}
}
\value{
\item{L}{value of the contrast}
}
\description{
Computation of the contrast based on the Euler approximation of -2 log-likelihood for the estimation
of the mixed SDE:

 \eqn{dXj(t)= (\alpha- \beta X_j(t))dt + \sigma_j a(X_j(t)) dW_j(t)}
 
with Gamma distribution for \eqn{1/\sigma_j^2} and fixed parameters in the drift.
}
\references{
Estimation of population parameters in stochastic differential equations with random effects in the diffusion coefficient, M. Delattre, V. Genon-Catalot and A. Samson, \emph{ESAIM: Probability and Statistics 2015}, Vol 19, \bold{671 -- 688}

Parametric inference for discrete observations of diffusion processes with mixed effects, M. Delattre, V. Genon-Catalot and C. Laredo, \emph{Preprint 2016, hal-01332630}
}
