% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastNormal.R
\name{contrastNormal}
\alias{contrastNormal}
\title{Computation of the contrast used for the estimation of the normal conditional distribution
of the random effects in the drift in mixed SDE with random effects in the drift and in the
diffusion coefficient}
\usage{
contrastNormal(mu, omega, U, V, S, K, estimphi, drift.random)
}
\arguments{
\item{mu}{value of the mean of the Gaussian distribution.}

\item{omega}{value of the standard deviation of the Gaussian distribution.}

\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{estimphi}{matrix of the M x 2 estimated parameters (\eqn{\alpha_j,\beta_j}).}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}
}
\value{
\item{L}{value of the contrast}
}
\description{
Computation of the contrast used for the estimation of the parameters of the 
Gaussian conditional distribution of the random effects in the drift \eqn{\alpha_j,\beta_j} when
the SDE includes random effects in the drift and in the diffusion coefficient:

 \eqn{dXj(t)= (\alpha_j - \beta_j Xj(t))dt + \sigma_j a(Xj(t)) dWj(t)}.
}
\references{
Estimaton of the joint distribution of random effects for a discretely observed diffusion with random effects, M. Delattre, V. Genon-Catalot and C. Laredo, \emph{Preprint, hal-01446063}.
}
