% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.randomvariable.R
\name{discr}
\alias{discr}
\title{Simulation Of Random Variables}
\usage{
discr(x, p)
}
\arguments{
\item{x}{n real numbers}

\item{p}{vector of probability, length n}
}
\value{
y a simulated value from the discrete distribution
}
\description{
Simulation of (discrete) random variables from a vector of probability (the nonparametrically estimated values
 of the density renormalised to sum at 1) and a vector of real values (the grid of estimation)
}
