% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstParamGamma.R
\name{EstParamGamma}
\alias{EstParamGamma}
\title{Estimation In Mixed Stochastic Differential Equations with fixed effects in the drift and one random effect in the diffusion coefficient}
\usage{
EstParamGamma(U, V, S, SigDelta, K, drift.param = c(0, 0), drift.estim = 1)
}
\arguments{
\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{SigDelta}{vector of the M constant terms of the individual likelihood (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{drift.param}{values of the fixed effects in the drift. Defaults to c(0,0).}

\item{drift.estim}{1 if the fixed effects in the drift are estimated, 0 otherwise. Defaults to 1.}
}
\value{
\item{mu}{values of the fixed effects in the drift.}
\item{a}{estimated value of the shape of the Gamma distribution.}
\item{lambda}{estimated value of the scale of the Gamma distribution.}
\item{BIChere}{BIC indicator.}
\item{AIChere}{AIC indicator.}
}
\description{
Parameter estimation of the mixed SDE with Gamma distribution of the diffusion random effect 
and fixed effects in the drift:

 \eqn{dXj(t)= (\alpha- \beta Xj(t))dt + \sigma_j a(Xj(t)) dWj(t), 1/\sigma_j^2 \sim \Gamma(a,lambda)}, 
 
done with \code{\link{likelihoodGamma}}.
}
\references{
Estimation of population parameters in stochastic differential equations with random effects in the diffusion coefficient, M. Delattre, V. Genon-Catalot and A. Samson, \emph{ESAIM: Probability and Statistics 2015}, Vol 19, \bold{671 -- 688}
}
