% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstParamNormalGamma.R
\name{EstParamNormalGamma}
\alias{EstParamNormalGamma}
\title{Estimation In Mixed Stochastic Differential Equations with random effects in the drift and in the diffusion coefficient}
\usage{
EstParamNormalGamma(U, V, S, SigDelta, K, drift.random, drift.fixed = 0,
  estim.drift.fix = 0)
}
\arguments{
\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{SigDelta}{vector of the M constant terms of the individual likelihood (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}

\item{drift.fixed}{value of the fixed effect in the drift if it is not estimated. Default to 0.}

\item{estim.drift.fix}{1 if the fixed effect in the drift is estimated, 0 otherwise. Default to 0.}
}
\value{
\item{mu}{estimated value of the mean of the Normal distribution}
\item{omega}{estimated value of the standard deviation of the Normal distribution}
\item{a}{estimated value of the shape of the Gamma distribution.}
\item{lambda}{estimated value of the scale of the Gamma distribution.}
\item{BIChere}{BIC indicator}
\item{AIChere}{AIC indicator}
}
\description{
Estimation of the parameters of the mixed SDE with Normal distribution of the random effects in the drift and
the square root of an inverse gamma distributed random effect in the diffusion:
 \eqn{dXj(t)= (\alpha- \beta Xj(t))dt + \sigma a(Xj(t)) dWj(t)}.
}
\references{
Estimaton of the joint distribution of random effects for a discretely observed diffusion with random effects, M. Delattre, V. Genon-Catalot and C. Laredo, \emph{Preprint}, hal-01446063.
}
