% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{class}
\name{Freq.fit-class}
\alias{Freq.fit-class}
\title{S4 class for the parametric estimation results}
\description{
S4 class for the parametric estimation results
}
\section{Slots}{

\describe{
\item{\code{model}}{character 'OU' or 'CIR'}

\item{\code{drift.random}}{numeric 1, 2, or c(1,2)}

\item{\code{diffusion.random}}{numeric 0 or 1}

\item{\code{gridf}}{matrix of values on which the estimation of the density of the random effects in the drift is done}

\item{\code{gridg}}{matrix of values on which the estimation of the density of the random effects in the diffusion is done}

\item{\code{mu}}{numeric MLE estimator for parametric approach}

\item{\code{omega}}{numeric  MLE estimator for parametric approach}

\item{\code{a}}{numeric MLE estimator for parametric approach}

\item{\code{lambda}}{numeric MLE estimator for parametric approach}

\item{\code{cutoff}}{value of the cutoff if there is one}

\item{\code{sigma2}}{numeric estimated value of \eqn{\sigma^2} if the diffusion coefficient is not random}

\item{\code{index}}{index of the used trajectories}

\item{\code{estimphi}}{matrix of the estimator of the drift random effects}

\item{\code{estimpsi2}}{vector of the estimator of the diffusion random effects}

\item{\code{estimf}}{estimator of the (conditional) density of \eqn{\phi}, matrix form}

\item{\code{estimg}}{estimator of the density of \eqn{\phi}, matrix form}

\item{\code{estim.drift.fix}}{1 if the user asked for the estimation of fixed parameter in the drift}

\item{\code{estim.diffusion.fix}}{1 if the user asked for the estimation of fixed diffusion coefficient}

\item{\code{discrete}}{1 if the estimation is based on the likelihood of discrete observations, 0 otherwise}

\item{\code{bic}}{numeric bic}

\item{\code{aic}}{numeric aic}

\item{\code{times}}{vector of observation times, storage of input variable}

\item{\code{X}}{matrix of observations, storage of input variable}
}}

