% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodGamma.R
\name{likelihoodGamma}
\alias{likelihoodGamma}
\title{Computation of the Euler approximation of -2 log-likelihood when there is one random 
effect in the diffusion coefficient.}
\usage{
likelihoodGamma(a, lambda, U, V, S, SigDelta, K, drift.fixed)
}
\arguments{
\item{a}{value of the shape of the Gamma distribution.}

\item{lambda}{value of the scape of the Gamma distribution.}

\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{SigDelta}{vector of the M constant terms of the individual likelihood (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{drift.fixed}{values of the fixed effects in the drift.}
}
\value{
\item{L}{value of -2 x loglikelihood}
}
\description{
Computation of the Euler approximation of -2 loglikelihood of the mixed SDE:
 \eqn{dXj(t)= (\alpha- \beta Xj(t))dt + \sigma_j a(Xj(t)) dWj(t)}
with Gamma distribution for \eqn{1/\sigma_j^2} and fixed parameters in the drift.
}
\references{
Estimation of population parameters in stochastic differential equations with random effects in the diffusion coefficient, M. Delattre, V. Genon-Catalot and A. Samson, \emph{ESAIM: Probability and Statistics 2015}, Vol 19, \bold{671 -- 688}
Parametric inference for discrete observations of diffusion processes with mixed effects, M. Delattre, V. Genon-Catalot and C. Laredo, \emph{Preprint 2016}, hal-01332630
}
