% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{likelihoodMixtureNormal}
\alias{likelihoodMixtureNormal}
\title{Computation of the Log Likelihood In Mixed Stochastic Differential Equations}
\usage{
likelihoodMixtureNormal(mu, omega, sigma, mixt.prop, U, V, S, K, estimphi,
  drift.random)
}
\arguments{
\item{mu}{mean of the random effects. N x 2 matrix, first (resp. second) column is the mean of \eqn{\alpha_j} (resp. 
\eqn{\beta_j}) in each mixture component if \eqn{\alpha_j} (resp. \eqn{\beta_j}) is random, the fixed effect value otherwise.}

\item{omega}{standard deviation of the random effects. N x 2 matrix, the components corresponding to a fixed effect should be 
set to 0.}

\item{sigma}{value of the diffusion parameter.}

\item{mixt.prop}{vector of mixture proportions.}

\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{estimphi}{matrix of the estimators of the random effects in the drift.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random 
effect or c(1,2) if 2 random effects.}
}
\value{
\item{L}{value of -2 x loglikelihood}
}
\description{
Computation of -2 loglikelihood the mixed SDE 
 \eqn{dXj(t)= (\alpha_j - \beta_j Xj(t))dt + \sigma a(Xj(t)) dWj(t)} with random effects
 in the drift following a mixture of Gaussian distributions.
}
\references{
Mixtures of stochastic differential equations with random effects: application to data clustering, M. Delattre, V. Genon-Catalot and A. Samson, \emph{Journal of Statistical Planning and Inference 2016}, Vol 173, \bold{109--124}
}
