% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodNormal.R
\name{likelihoodNormal}
\alias{likelihoodNormal}
\title{Computation Of The Log Likelihood In Mixed Stochastic Differential Equations.}
\usage{
likelihoodNormal(mu, omega, sigma, U, V, S, SigDelta = 0, K, estimphi,
  drift.random, discrete = 1)
}
\arguments{
\item{mu}{current value of the mean of the normal distribution.}

\item{omega}{current value of the standard deviation of the normal distribution.}

\item{sigma}{current value of the diffusion coefficient.}

\item{U}{vector of the M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{vector of the M sufficient statistics V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{SigDelta}{vector of the M constant terms of the individual likelihood (see \code{\link{UVS}}). 
Required only if discrete = 1. Defaults to 0.}

\item{K}{number of times of observations.}

\item{estimphi}{matrix of estimators of the random effects.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}

\item{discrete}{1 for discrete observations, 0 otherwise. If discrete = 0, the exact likelihood associated with continuous observations is 
discretized. If discrete = 1, the likelihood of the Euler scheme of the mixed SDE is computed. Defaults to 1.}
}
\value{
\item{L}{value of -2 x loglikelihood}
}
\description{
Computation of -2 loglikelihood of the mixed SDE with Normal distribution of the random effects
and fixed effect in the diffusion coefficient:
 \eqn{dXj(t)= (\alpha_j- \beta_j Xj(t))dt + \sigma a(Xj(t)) dWj(t)}.
}
\references{
Maximum likelihood estimation for stochastic differential equations with random effects, M. Delattre, V. Genon-Catalot and A. Samson, \emph{Scandinavian Journal of Statistics 2012}, Vol 40, \bold{322--343}
Parametric inference for discrete observations of diffusion processes with mixed effects, M. Delattre, V. Genon-Catalot and C. Laredo, \emph{Preprint}, hal-01332630
}
