% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{likelihoodNormalindi}
\alias{likelihoodNormalindi}
\title{Computation of the individual Log-Likelihoods in Mixed Stochastic Differential Equations}
\usage{
likelihoodNormalindi(mu, omega, sigma, Uj, Vj, Sj, K, estimphij, drift.random)
}
\arguments{
\item{mu}{vector of mean of the random effects. First (resp. second) value is the mean of \eqn{\alpha_j} (resp. 
\eqn{\beta_j}) if \eqn{\alpha_j} (resp. \eqn{\beta_j}) is random, the fixed effect value otherwise.}

\item{omega}{standard deviation of the random effects. The components corresponding to a fixed effect should be 
set to 0.}

\item{sigma}{value of the diffusion parameter.}

\item{Uj}{vector of the sufficient statistics U for individual j (see \code{\link{UVS}}).}

\item{Vj}{matrix of the sufficient statistics V for individual j (see \code{\link{UVS}}).}

\item{Sj}{value of the sufficient statistic S for individual j (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{estimphij}{vector of estimators of the random effects for individual j.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random 
effect or c(1,2) if 2 random effects.}
}
\value{
\item{L}{value of -2 x loglikelihood}
}
\description{
Computation of -2 log-likelihood of individual j in the mixed SDE with Normal distribution of the drift random effects
and fixed effect in the diffusion:

 \eqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma a(X_j(t)) dW_j(t)}.
}
\references{
Maximum likelihood estimation for stochastic differential equations with random effects, M. Delattre, V. Genon-Catalot and A. Samson, \emph{Scandinavian Journal of Statistics 2012}, Vol 40, \bold{322--343}
}
