% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.sim.R
\name{mixture.sim}
\alias{mixture.sim}
\title{Simulation Of A Mixture Of Normal Distributions}
\usage{
mixture.sim(M, param, prob)
}
\arguments{
\item{M}{number of simulated variables}

\item{param}{vector of parameters with the means and standard-deviations of the normal distributions}

\item{prob}{mixture components probabilities}
}
\value{
\item{Y}{vector of simulated variables}
}
\description{
Simulation of M random variables from a mixture of Gaussian distributions
}
\details{
If the distribution is \eqn{p1 N(\mu1,\sigma1^2) + (1-p1)N(\mu2, \sigma2^2)}

param=c(\eqn{\mu1, \sigma1, \mu2, \sigma2}) and prob=c(p1,1-p1)
}
