% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.pred.R
\name{msde.pred}
\alias{msde.pred}
\title{Prediction Of Mixed Stochastic Differential Equations Trajectories}
\usage{
msde.pred(times, X, model = c("OU", "CIR"), drift.random,
  drift.fixed = NULL, estim.drift.fix = 0, diffusion.random = 0,
  diffusion.fixed = NULL, estim.diffusion.fix = 0, mixture = 0,
  nb.mixt = 1, drift.fixed.mixed = 0, Niter = 10, discrete = 1,
  plot.pred = TRUE, level = 0.05, newwindow = FALSE)
}
\arguments{
\item{times}{vector of observation times}

\item{X}{matrix of the M trajectories (each row is a trajectory with as much columns as observations)}

\item{model}{name of the SDE: 'OU' (Ornstein-Uhlenbeck) or 'CIR' (Cox-Ingersoll-Ross)}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects}

\item{drift.fixed}{default NULL, fixed effect in the drift: value of the fixed effect when there is only one random effect and it is not estimated, NULL otherwise}

\item{estim.drift.fix}{default 0, 1 if the fixed effect in the drift is estimated, 0 if the fixed effect in the drift is known}

\item{diffusion.random}{default 0, 1 if one random effect in the diffusion, 0 if there is no random effect in the diffusion}

\item{diffusion.fixed}{default NULL, fixed effect in the diffusion: value of the fixed effect when there is no random effect in the diffusion and it is not estimated, NULL otherwise}

\item{estim.diffusion.fix}{default 0, 1 if the fixed effect in the diffusion is estimated, 0 otherwise}

\item{mixture}{1 if the random effects in the drift follow a mixture distribution, 0 otherwise. Default to 0.}

\item{nb.mixt}{default 1, number of mixture components for the distribution of the random effects in the drift}

\item{drift.fixed.mixed}{default 1, 1 if the value of the fixed effect in the drift is different from one mixture component to
another, 0 otherwise.}

\item{Niter}{default 10, number of iterations for the EM algorithm if mixture = 1}

\item{discrete}{default 1, 1 for discrete observations, 0 otherwise. If discrete = 0, and diffusion.random = 0, the exact likelihood associated with continuous observations is 
discretized. If discrete = 1, the likelihood of the Euler scheme of the mixed SDE is computed.}

\item{plot.pred}{logical(1), if TRUE, the results are depicted grafically}

\item{level}{alpha for the predicion intervals, default 0.05}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}
}
\value{
\item{res}{is the vector of subscript in \eqn{1,...,M} where the estimation of \eqn{phi} has been done,  most of the time \eqn{index= 1:M}}
\item{Xpred}{is the vector of subscript in \eqn{1,...,M} where the estimation of \eqn{phi} has been done,  most of the time \eqn{index= 1:M}}
\item{indexpred}{matrix of estimators of \eqn{\phi=\alpha, or \beta, or (\alpha,\beta)} from the efficient statitics (see \code{\link{UVS}}), matrix of two lines if drift.random =c(1,2), numerical type otherwise}
\item{phipred}{matrix of estimators of \eqn{\psi^2=\sigma^2} from the efficient statistics (see \code{\link{UVS}}), matrix of one line}
}
\description{
This function proposes to keep two thirds of the data (randomly chosen) to do the parametric
estimation of the density of the parameters and of the fixed parameters of model 
 \eqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma_j a(X_j(t)) dW_j(t)}
using the same method as in the estimationr function \code{msde.fit}
and then to predict new trajectories from the estimated model. 
The plot reflect the adequation between the last third of the data and the simulated one.
}
\examples{

\dontrun{
# Example 1: one random effect in the drift and one random effect in the diffusion coefficient.

# -- Simulation
M <- 100
Tmax <- 5
N <- 5000
model <- 'OU'
drift.random <- 2
diffusion.random <- 1
drift.fixed <- 0
drift.param <- c(0.5,0.5)
diffusion.param <- c(8,1/2)

sim1 <- msde.sim(M = M, T = Tmax, N = N, model = model, drift.random = drift.random, 
diffusion.random = diffusion.random, drift.fixed = drift.fixed, 
mixture = 0, drift.param = drift.param, diffusion.param = diffusion.param)
                 

pred1 <- pred( times = sim1$times, X = sim1$X, model = model, drift.random = drift.random,
              estim.drift.fix = estim.drift.fix, diffusion.random = diffusion.random, 
              estim.diffusion.fix = estim.diffusion.fix,level = 0.05)



# Example 5: one fixed effect and one mixture random effect in the drift, and one fixed effect in 
# the diffusion coefficient

# -- Simulation
M <- 100
Tmax <- 5
N <- 5000
diffusion.random <- 0
diffusion.param <- 0.1
model <- 'OU'
drift.random <- 1
drift.fixed <- 1
nb.mixt <- 2
mixt.prop <- c(0.5,0.5)
param.ea1 <- c(0.5, 0.25, 1.8, 0.25)
param.ea2 <- c(1, 0.25, 1, 0.25) 
drift.param <- param.ea1

sim5 <- msde.sim(M = M, T = Tmax, N = N, model = model, drift.random = drift.random,
                 diffusion.random = diffusion.random, drift.fixed = drift.fixed,
                 mixture=1, drift.param = drift.param,
                 diffusion.param = diffusion.param, nb.mixt = nb.mixt, mixt.prop = mixt.prop)


res5 <- msde.pred(times = sim5$times, X = sim5$X, model = model, drift.random = drift.random,
                  estim.drift.fix = estim.drift.fix, diffusion.random = diffusion.random, estim.diffusion.fix = estim.diffusion.fix,
                  mixture = mixture, nb.mixt=nb.mixt, Niter = Niter)

  }

}
\references{
See  
Maximum Likelihood Estimation for Stochastic Differential Equations with Random Effects, Delattre, M., Genon-Catalot, V. and Samson, A. \emph{Scandinavian Journal of Statistics 40(2) 2012} \bold{322-343} 

Estimation of population parameters in stochastic differential equations with random effects in the diffusion coefficient, Delattre, M., Genon-Catalot, V. and Samson, A. \emph{ESAIM:PS 19 2015} \bold{671-688}

Mixtures of stochastic differential equations with random effects: application to data clustering, Delattre, M., Genon-Catalot, V. and Samson, A. \emph{Journal of Statistical Planning and Inference 173 2016} \bold{109-124} 

Parametric inference for discrete observations of diffusion processes with mixed effects, Delattre, M., Genon-Catalot, V. and Laredo, C. \emph{hal-01332630 2016}

Estimation of the joint distribution of random effects for a discretely observed diffusion with random effects, Delattre, M., Genon-Catalot, V. and Laredo, C. \emph{hal-01446063 2017}
}
\keyword{estimation}
