% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.sim.R
\name{msde.sim}
\alias{msde.sim}
\title{Simulation Of A Mixed Stochastic Differential Equation}
\usage{
msde.sim(M, T, N = 100, model, drift.random, diffusion.random, density.phi,
  drift.fixed = 0, drift.param, diffusion.param, nb.mixt = 1,
  mixt.prop = 1, t0 = 0, X0 = 0.01, invariant = 0, delta = T/N,
  op.plot = 0, add.plot = FALSE)
}
\arguments{
\item{M}{number of trajectories.}

\item{T}{horizon of simulation.}

\item{N}{number of simulation steps, default Tx100.}

\item{model}{name of the SDE: 'OU' (Ornstein-Uhlenbeck) or 'CIR' (Cox-Ingersoll-Ross).}

\item{drift.random}{random effects in the drift: 0 if no random effect, 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}

\item{diffusion.random}{random effect in the diffusion coefficient: 0 if no random effect, 1 if one multiplicative random effect.}

\item{density.phi}{name of the density of the random effects.}

\item{drift.fixed}{fixed effects in the drift: value of the fixed effect when there is only one random effect, 0 otherwise. 
If drift.random =2, fixed can be 0 but \eqn{\beta} has to be a non negative random variable for the estimation.}

\item{drift.param}{vector of parameters of the distribution of the random effects in the drift.}

\item{diffusion.param}{diffusion parameter if the diffusion coefficient is fixed, vector of parameters of the distribution of the diffusion random effect otherwise.}

\item{nb.mixt}{number of mixture components if the drift random effects follow a mixture distribution, default nb.mixt=1.}

\item{mixt.prop}{vector of mixture proportions if the drift random effects follow a mixture distribution, default mixt.prop=1.}

\item{t0}{time origin, default 0.}

\item{X0}{initial value of the process, default X0=0.}

\item{invariant}{1 if the initial value is simulated from the invariant distribution, default 0.01 and X0 is fixed.}

\item{delta}{time step of the simulation (T/N).}

\item{op.plot}{1 if a plot of the trajectories is required, default 0.}

\item{add.plot}{1 for add trajectories to an existing plot}
}
\value{
\item{X}{matrix (M x (N+1)) of the M trajectories. }
\item{phi}{vector (or matrix) of the M simulated random effects.}
}
\description{
Simulation of M independent trajectories of a mixed stochastic differential equation (SDE) with linear drift  and two random effects \eqn{(\alpha_j, \beta_j)}
 \deqn{dX_j(t)= (\alpha_j- \beta_j X_(t))dt + \sigma a(X_j(t)) dW_j(t)  ,  j=1, ..., M.}
}
\details{
Simulation of M independent trajectories of the SDE (the Brownian motions \eqn{Wj} are independent), with linear drift. Two diffusions are implemented, with one or two random effects:
\subsection{Ornstein-Uhlenbeck model (OU)}{
If random = 1, \eqn{\beta} is a fixed effect: \eqn{dX_j(t)= (\alpha_j- \beta X_j(t))dt + \sigma_j dW_j(t)  } 

If random = 2, \eqn{\alpha} is a fixed effect: \eqn{dX_j(t)= (\alpha - \beta_j X_j(t))dt + \sigma_j dW_j(t)  }

If random = c(1,2), \eqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma_j dW_j(t)  } 
}
\subsection{Cox-Ingersoll-Ross model (CIR)}{
If random = 1, \eqn{\beta} is a fixed effect: \eqn{dX_j(t)= (\alpha_j- \beta X_j(t))dt + \sigma_j \sqrt{X_j(t)} dW_j(t)  } 

If random = 2, \eqn{\alpha} is a fixed effect: \eqn{dX_j(t)= (\alpha - \beta_j X_j(t))dt + \sigma_j \sqrt{X_j(t)} dW_j(t)  } 

If random = c(1,2), \eqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma_j \sqrt{X_j(t)}  dW_j(t)  } 
}
The initial value of each trajectory can be simulated from the invariant distribution of the process:
 Normal distribution with mean \eqn{\alpha/\beta} and variance \eqn{\sigma^2/(2 \beta)} for the OU,  a gamma distribution
 \eqn{\Gamma(2\alpha/\sigma^2,  \sigma^2/(2\beta))} for the C-I-R model.

\subsection{Density of the diffusion random effect}{
We only consider inverse Gamma distributions for \eqn{\sigma_j^2} and diffusion.param=c(shape,scale). 

}

\subsection{Density of the drift random effects}{
Several densities are implemented for the random effects, depending on the number of random effects. 

\emph{If two random effects, choice between} 

'normixtnormixt': Mixture of \eqn{N} normal distributions for both  \eqn{\alpha} \eqn{\beta} and param=c(mean_\eqn{\alpha}^\eqn{1}, sd_\eqn{\alpha}^\eqn{1}, ..., mean_\eqn{\alpha}^\eqn{N}, sd_\eqn{\alpha}^\eqn{N},mean_\eqn{\beta}^\eqn{1}, sd_\eqn{\beta}^\eqn{1}, ..., mean_\eqn{\beta}^\eqn{N}, sd_\eqn{\beta}^\eqn{N})

'normalnormal': Normal distributions for both  \eqn{\alpha} \eqn{\beta} and param=c(mean_\eqn{\alpha}, sd_\eqn{\alpha}, mean_\eqn{\beta}, sd_\eqn{\beta})

'gammagamma': Gamma distributions for both  \eqn{\alpha} \eqn{\beta} and param=c(shape_\eqn{\alpha}, scale_\eqn{\alpha}, shape_\eqn{\beta}, scale_\eqn{\beta})

'gammainvgamma': Gamma for \eqn{\alpha}, Inverse Gamma for \eqn{\beta} and param=c(shape_\eqn{\alpha}, scale_\eqn{\alpha}, shape_\eqn{\beta}, scale_\eqn{\beta})

'normalgamma':  Normal for \eqn{\alpha}, Gamma for \eqn{\beta} and param=c(mean_\eqn{\alpha}, sd_\eqn{\alpha}, shape_\eqn{\beta}, scale_\eqn{\beta})

'normalinvgamma':  Normal for \eqn{\alpha}, Inverse Gamma for \eqn{\beta} and param=c(mean_\eqn{\alpha}, sd_\eqn{\alpha}, shape_\eqn{\beta}, scale_\eqn{\beta})

'gammagamma2':  Gamma \eqn{+2 * \sigma^2} for \eqn{\alpha},  Gamma \eqn{+ 1} for \eqn{\beta}  and param=c(shape_\eqn{\alpha}, scale_\eqn{\alpha}, shape_\eqn{\beta}, scale_\eqn{\beta})

'gammainvgamma2':   Gamma \eqn{+2 * \sigma^2} for \eqn{\alpha}, Inverse Gamma for \eqn{\beta} and param=c(shape_\eqn{\alpha}, scale_\eqn{\alpha}, shape_\eqn{\beta}, scale_\eqn{\beta})

\emph{If only \eqn{\alpha} is random, choice between}

'normixt': Mixture of \eqn{N} normal distributions and param=c(mean_\eqn{\alpha}^\eqn{1}, sd_\eqn{\alpha}^\eqn{1}, ..., mean_\eqn{\alpha}^\eqn{N}, sd_\eqn{\alpha}^\eqn{N})

'normal': Normal distribution with param=c(mean, sd)

lognormal': logNormal distribution with param=c(mean, sd)

'mixture.normal': mixture of normal distributions \eqn{p N(\mu1,\sigma1^2) + (1-p)N(\mu2, \sigma2^2)} with 
param=c(p, \eqn{\mu1, \sigma1, \mu2, \sigma2})

'gamma': Gamma distribution with param=c(shape, scale) 

'mixture.gamma': mixture of Gamma distribution \eqn{p \Gamma(shape1,scale1) + (1-p)\Gamma(shape2,scale2)}
with param=c(p, shape1, scale1, shape2, scale2)

'gamma2': Gamma distribution \eqn{+2 * \sigma^2} with param=c(shape, scale) 

'mixed.gamma2': mixture of Gamma distribution \eqn{p \Gamma(shape1,scale1) + (1-p) \Gamma(shape2,scale2)} + \eqn{+2 * \sigma^2}
with param=c(p, shape1, scale1, shape2, scale2)

\emph{If only \eqn{\beta} is random, choice between}

 'normixt': Mixture of \eqn{N} normal distributions and param=c(mean_\eqn{\beta}^\eqn{1}, sd_\eqn{\beta}^\eqn{1}, ..., mean_\eqn{\beta}^\eqn{N}, sd_\eqn{\beta}^\eqn{N})

 'normal': Normal distribution with param=c(mean, sd)
 
 'gamma': Gamma distribution with param=c(shape, scale) 
 
 'mixture.gamma': mixture of Gamma distribution \eqn{p \Gamma(shape1,scale1) + (1-p) \Gamma(shape2,scale2)}
with param=c(p, shape1, scale1, shape2, scale2)

}
}
\references{
This function mixedsde.sim is based on the package sde, function sde.sim. See Simulation and Inference for stochastic differential equation, S.Iacus, \emph{Springer Series in Statistics 2008}
Chapter 2
}
\seealso{
\url{http://cran.r-project.org/package=sde}
}
