\name{MsdeParEst-package}
\alias{MsdeParEst-package}
\alias{MsdeParEst}
\docType{package}
\title{
\packageTitle{MsdeParEst}
}
\description{
\packageDescription{MsdeParEst}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MsdeParEst}
\packageIndices{MsdeParEst}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MsdeParEst}

Maintainer: \packageMaintainer{MsdeParEst}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
