% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{plot,Fit.class,ANY-method}
\alias{plot,Fit.class,ANY-method}
\title{Plot method for the estimation class object}
\usage{
\S4method{plot}{Fit.class,ANY}(x, newwindow = FALSE, ...)
}
\arguments{
\item{x}{Fit.class class}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{...}{optional plot parameters}
}
\description{
Plot method for the S4 class Fit.class
}
