% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{plot,Freq.mixture.fit,ANY-method}
\alias{plot,Freq.mixture.fit,ANY-method}
\title{Plot method for the frequentist estimation class object}
\usage{
\S4method{plot}{Freq.mixture.fit,ANY}(x, newwindow = FALSE, ...)
}
\arguments{
\item{x}{Freq.mixture.fit class}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{...}{optional plot parameters}
}
\description{
Plot method for the S4 class Freq.mixture.fit
}
