% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{pred,Freq.fit-method}
\alias{pred,Freq.fit-method}
\title{Prediction method for the Freq.fit class object}
\usage{
\S4method{pred}{Freq.fit}(x, invariant = 0, level = 0.05,
  newwindow = FALSE, plot.pred = TRUE, ...)
}
\arguments{
\item{x}{Freq.fit class}

\item{invariant}{1 if the initial value is from the invariant distribution, default X0 is fixed from Xtrue}

\item{level}{alpha for the predicion intervals, default 0.05}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.pred}{logical(1), if TRUE, the results are depicted grafically}

\item{...}{optional plot parameters}
}
\description{
Frequentist prediction
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}
