% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{print,Freq.mixture.fit-method}
\alias{print,Freq.mixture.fit-method}
\title{Description of print}
\usage{
\S4method{print}{Freq.mixture.fit}(x)
}
\arguments{
\item{x}{Freq.mixture.fit class}
}
\description{
Method for the S4 class Freq.mixture.fit
}
