% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{summary,Fit.class-method}
\alias{summary,Fit.class-method}
\title{Short summary of the results of class object Fit.class}
\usage{
\S4method{summary}{Fit.class}(object)
}
\arguments{
\item{object}{Fit.class class}
}
\description{
Method for the S4 class Fit.class
}
