% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{summary,Freq.mixture.fit-method}
\alias{summary,Freq.mixture.fit-method}
\title{Short summary of the results of class object Freq.mixture.fit}
\usage{
\S4method{summary}{Freq.mixture.fit}(object)
}
\arguments{
\item{object}{Freq.mixture.fit class}
}
\description{
Method for the S4 class Freq.fit
}
