% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{summary,Mixture.fit.class-method}
\alias{summary,Mixture.fit.class-method}
\title{Short summary of the results of class object Mixture.fit.class}
\usage{
\S4method{summary}{Mixture.fit.class}(object)
}
\arguments{
\item{object}{Mixture.fit.class class}
}
\description{
Method for the S4 class Mixture.fit.class
}
