% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVS.R
\name{UVS}
\alias{UVS}
\title{Computation Of The Sufficient Statistics}
\usage{
UVS(X, model, times)
}
\arguments{
\item{X}{matrix of the M trajectories.}

\item{model}{name of the SDE: 'OU' (Ornstein-Uhlenbeck) or 'CIR' (Cox-Ingersoll-Ross).}

\item{times}{times vector of observation times.}
}
\value{
\item{U}{vector of the M statistics U(Tend)}
\item{V}{list of the M matrices V(Tend)}
\item{S}{vector of the M quadratic variations S(X(t_1),...,X(t_n))}
\item{SigDelta}{vector of the M constant contributions to the Euler scheme approximation to the likelihood SigDelta(X(t_1),...,X(t_n))}
}
\description{
Computation of the sufficient statistics of the (approximate) likelihood of the mixed SDE
 \eqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma_j a(X_j(t)) dW_j(t)}.
}
\details{
Computation of the sufficient statistics of the (approximate) likelihood of the mixed SDE
\eqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma_j a(X_j(t)) dW_j(t) = (\alpha_j, \beta_j)b(X_j(t))dt + \sigma_j a(X_j(t)) dW_j(t)} with \eqn{b(x)=(1,-x)^t}:

U : \eqn{U(Tend) = \int_0^{Tend} b(X(s))/a^2(X(s))dX(s) }
 
V : \eqn{V(Tend) = \int_0^{Tend} b(X(s))^2/a^2(X(s))ds }

S : \eqn{S(X(t_1),...,X(t_n)) = 1/delta \sum_{j=1}^{n} (X(t_j)-X(t_{j-1}))^2/a^2(X(t_{j-1}))}
 
SigDelta: \eqn{SigDelta(X(t_1),...,X(t_n)) = n log(delta) + \sum_{j=1}^{n} log(a(X(t_j)))}
}
\references{
See  
Maximum Likelihood Estimation for Stochastic Differential Equations with Random Effects, Delattre, M., Genon-Catalot, V. and Samson, A. \emph{Scandinavian Journal of Statistics 40(2) 2012} \bold{322-343} 
Estimation of population parameters in stochastic differential equations with random effects in the diffusion coefficient, Delattre, M., Genon-Catalot, V. and Samson, A. \emph{ESAIM:PS 19 2015} \bold{671-688}
Mixtures of stochastic differential equations with random effects: application to data clustering, Delattre, M., Genon-Catalot, V. and Samson, A. \emph{Journal of Statistical Planning and Inference 173 2016} \bold{109-124} 
Parametric inference for discrete observations of diffusion processes with mixed effects, Delattre, M., Genon-Catalot, V. and Laredo, C. \emph{hal-01332630 2016}
Estimation of the joint distribution of random effects for a discretely observed diffusion with random effects, Delattre, M., Genon-Catalot, V. and Laredo, C. \emph{hal-01446063 2017}
}
\keyword{quadratic}
\keyword{statistics,}
\keyword{sufficient}
\keyword{variations}
