% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.fit.R
\docType{methods}
\name{valid,Fit.class-method}
\alias{valid,Fit.class-method}
\title{Validation of the chosen model.}
\usage{
\S4method{valid}{Fit.class}(x, Mrep = 100, newwindow = FALSE,
  plot.valid = TRUE, numj, ...)
}
\arguments{
\item{x}{Fit.class class}

\item{Mrep}{number of trajectories to be drawn}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.valid}{logical(1), if TRUE, the results are depicted grafically}

\item{numj}{optional number of series to be validated}

\item{...}{optional plot parameters}
}
\description{
Validation of the chosen model. For the index numj, Mrep=100 new trajectories are simulated
with the value of the estimated random effect number numj. Two plots are given: on the left the simulated trajectories and the true one (red)
and one the left the corresponding qq-plot for each time.
}
