% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx.R
\name{bx}
\alias{bx}
\title{Computation Of The Drift Coefficient}
\usage{
bx(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
\item{b}{The drift is \eqn{b(x,\phi) = \phi_1 b_1(x) + \phi_2 b_2(x)}, the output is the vector \eqn{(b_1,b_2)^t}}
}
\description{
Computation of the drift coefficient
}
\keyword{drift}
