% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.pred.R
\docType{class}
\name{class.mixture.pred-class}
\alias{class.mixture.pred-class}
\title{S4 class for the parametric estimation results when the random effects in the drift follow
mixture of normal distributions}
\description{
S4 class for the parametric estimation results when the random effects in the drift follow
mixture of normal distributions
}
\section{Slots}{

\describe{
\item{\code{estim}}{object of class Mixture.fit.class containing the results of the model estimation}

\item{\code{phipred}}{numeric 1, 2, or c(1,2)}

\item{\code{Xpred}}{matrix of predicted trajectories (dimensions)}

\item{\code{idexpred}}{matrix of values on which the estimation of the density of the random effects is done}
}}

