% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.pred.R
\docType{class}
\name{class.pred-class}
\alias{class.pred-class}
\title{S4 class for the estimation results in the mixed SDE with random effects in the drift, in the diffusion or both}
\description{
S4 class for the estimation results in the mixed SDE with random effects in the drift, in the diffusion or both
}
\section{Slots}{

\describe{
\item{\code{res}}{character 'OU' or 'CIR'}

\item{\code{phipred}}{numeric 1, 2, or c(1,2)}

\item{\code{idexpred}}{matrix of values on which the estimation of the density of the random effects is done}
}}

