% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM algorithm for mixtures of stochastic differential equations with random effects in the drift}
\usage{
EM(U, V, S, K, drift.random, start, Niter = 10, drift.fixed = 0,
  drift.estim.fixed = 1, drift.fixed.mixt = 1, sigma = 1,
  sigma.estim = 1)
}
\arguments{
\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{K}{number of times of observations.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random 
effect or c(1,2) if 2 random effects.}

\item{start}{list of starting values: mu, omega, mixt.prop. mixt.prop is a vector of length N, the number of 
mixture components. mu is a N x 2 matrix, first (resp. second) column is the mean of \eqn{\alpha_j} (resp. 
\eqn{\beta_j}) if \eqn{\alpha_j} (resp. \eqn{\beta_j}) is random, the fixed effect value otherwise. omega is a 
N x 2 matrix, the components corresponding to a fixed effect should be set to 0.}

\item{Niter}{number of iterations. Defaults to 10.}

\item{drift.fixed}{value for the fixed effects in the drift if known (not estimated). Vector of length N. Defaults to 0.}

\item{drift.estim.fixed}{1 if the fixed effects in the drift are estimated, 0 otherwise. Defaults to 1.}

\item{drift.fixed.mixt}{1 if the value of the fixed effect in the drift is different from one mixture component to
another, 0 otherwise. Defaults to 1.}

\item{sigma}{value for the diffusion parameter if known (not estimated). Defaults to 1.}

\item{sigma.estim}{1 if the diffusion parameter is estimated, 0 otherwise. Defaults to 1.}
}
\value{
\item{mu}{estimated value of the mean at each iteration of the algorithm. Niter x N x 2 array. }
\item{omega}{estimated value of the standard deviation at each iteration of the algorithm. Niter x N x 2 array.}
\item{mixt.prop}{estimated value of the mixture proportions at each iteration of the algorithm. Niter x N matrix.}
\item{sigma}{value of the diffusion parameter.}
\item{probindi}{posterior component probabilites. M x N matrix.}
\item{BIChere}{BIC indicator}
\item{AIChere}{AIC indicator}
}
\description{
EM algorithm for parameter estimation in the mixed SDE 
 \eqn{dXj(t)= (\alpha_j - \beta_j Xj(t))dt + \sigma a(Xj(t)) dWj(t)} with random effects
 in the drift following a mixture of Gaussian distributions.
}
\references{
Mixtures of stochastic differential equations with random effects: application to data clustering, M. Delattre, V. Genon-Catalot and A. Samson, \emph{Journal of Statistical Planning and Inference 2016}, Vol 173, \bold{109--124}
}
