% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstParamNormal.R
\name{EstParamNormal}
\alias{EstParamNormal}
\title{Estimation In Mixed Stochastic Differential Equations with random effects in the drift and fixed effect in the diffusion coefficient}
\usage{
EstParamNormal(U, V, S, SigDelta = 0, K, drift.fixed = 0,
  estim.drift.fix = 1, sigma = 0, drift.random, diffusion.estim = 1,
  discrete = 1)
}
\arguments{
\item{U}{matrix of M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{list of the M sufficient statistics matrix V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{SigDelta}{vector of the M constant terms of the individual likelihood (see \code{\link{UVS}}). 
Required only if discrete = 1. Defaults to 0.}

\item{K}{number of times of observations.}

\item{drift.fixed}{value of the fixed effect in the drift if it is not estimated. Default to 0.}

\item{estim.drift.fix}{1 if the fixed effect in the drift is estimated, 0 otherwise. Default to 1.}

\item{sigma}{value of the fixed effect in the diffusion if known (not estimated). Defaults to 0.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}

\item{diffusion.estim}{1 if sigma is estimated, 0 otherwise.}

\item{discrete}{1 for discrete observations, 0 otherwise. If discrete = 0, the exact likelihood associated with continuous observations is 
discretized. If discrete = 1, the likelihood of the Euler scheme of the mixed SDE is computed. Defaults to 1.}
}
\value{
\item{mu}{estimated value of the mean of the Normal distribution}
\item{omega}{estimated value of the standard deviation of the Normal distribution}
\item{sigma}{value of the diffusion coefficient}
\item{BIChere}{BIC indicator}
\item{AIChere}{AIC indicator}
}
\description{
Estimation of the parameters of the mixed SDE with Normal distribution of the random effects in the drift and
fixed parameter in the diffusion:
 \eqn{dXj(t)= (\alpha_j - \beta_j Xj(t))dt + \sigma a(Xj(t)) dWj(t)}, done with \code{\link{likelihoodNormal}}.
}
