% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodNormalGamma.R
\name{likelihoodNormalGamma}
\alias{likelihoodNormalGamma}
\title{Computation Of The Log Likelihood of the Euer scheme in Mixed Stochastic Differential Equations.}
\usage{
likelihoodNormalGamma(a, lambda, mu, omega, U, V, S, SigDelta, K, drift.random)
}
\arguments{
\item{a}{current value of the shape of the Gamma distribution.}

\item{lambda}{current value of the scape of the Gamma distribution.}

\item{mu}{current value of the mean of the normal distribution.}

\item{omega}{current value of the standard deviation of the normal distribution.}

\item{U}{vector of the M sufficient statistics U (see \code{\link{UVS}}).}

\item{V}{vector of the M sufficient statistics V (see \code{\link{UVS}}).}

\item{S}{vector of the M sufficient statistics S (see \code{\link{UVS}}).}

\item{SigDelta}{vector of the M constant terms of the individual likelihood (see \code{\link{UVS}}). 
Required only if discrete = 1. Defaults to 0.}

\item{K}{number of times of observations.}

\item{drift.random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}
}
\value{
\item{L}{value of -2 x loglikelihood}
}
\description{
Computation of -2 loglikelihood of the Euler scheme the mixed SDE 
and fixed effect in the diffusion coefficient:
 \eqn{dXj(t)= (\alpha_j- \beta_j Xj(t))dt + \sigma_j a(Xj(t)) dWj(t)},
with Normal conditional distribution of the random effects in the drift and Gamma distribution of 
\eqn{1/\sigma_j^2} in the diffusion.
}
\references{
Estimaton of the joint distribution of random effects for a discretely observed diffusion with random effects, M. Delattre, V. Genon-Catalot and C. Laredo, \emph{Preprint}, hal-01446063.
}
