% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.sim.R
\name{mixture.sim}
\alias{mixture.sim}
\title{Simulation Of A Mixture Of Two Normal Or Gamma Distributions}
\usage{
mixture.sim(M, density.phi, param, prob)
}
\arguments{
\item{M}{number of simulated variables}

\item{density.phi}{name of the chosen density 'mixture.normal' or 'mixture.gamma'}

\item{param}{vector of parameters with the proportion of mixture of the two distributions and means and standard-deviations of the two normal or 
shapes and scales of the two Gamma distribution}

\item{prob}{mixture components probabilities}
}
\value{
\item{Y}{vector of simulated variables}
}
\description{
Simulation of M random variables from a mixture of two Gaussian or Gamma distributions
}
\details{
If 'mixture.normal', the distribution is \eqn{p N(\mu1,\sigma1^2) + (1-p)N(\mu2, \sigma2^2)}

and param=c(p, \eqn{\mu1, \sigma1, \mu2, \sigma2})

If 'mixture.gamma', the distribution is \eqn{p Gamma(shape1,scale1) + (1-p)Gamma(shape2,scale2)}

and param=c(p, shape1, scale1, shape2, scale2)
}
