% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msde.sim.R
\name{msde.sim}
\alias{msde.sim}
\title{Simulation Of A Mixed Stochastic Differential Equation}
\usage{
msde.sim(M, T, N = 100, model, drift.random, diffusion.random, mixture = 0,
  drift.fixed = 0, drift.param, diffusion.param, nb.mixt = 1,
  mixt.prop = 1, t0 = 0, X0 = 0.01, invariant = 0, delta = T/N,
  op.plot = 0, add.plot = FALSE)
}
\arguments{
\item{M}{number of trajectories.}

\item{T}{horizon of simulation.}

\item{N}{number of simulation steps, default Tx100.}

\item{model}{name of the SDE: 'OU' (Ornstein-Uhlenbeck) or 'CIR' (Cox-Ingersoll-Ross).}

\item{drift.random}{random effects in the drift: 0 if no random effect, 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}

\item{diffusion.random}{random effect in the diffusion coefficient: 0 if no random effect, 1 if one multiplicative random effect.}

\item{mixture}{1 if the random effects in the drift follow a mixture of Normal distributions, 0 otherwise. Default to 0.}

\item{drift.fixed}{fixed effects in the drift: value of the fixed effect when there is only one random effect, 0 otherwise. 
If drift.random =2, fixed can be 0 but \eqn{\beta} has to be a non negative random variable for the estimation.}

\item{drift.param}{vector (not mixture) or matrix (mixture) of parameters of the distribution of the random effects in the drift.}

\item{diffusion.param}{diffusion parameter if the diffusion coefficient is fixed, vector of parameters of the distribution of the diffusion random effect otherwise.}

\item{nb.mixt}{number of mixture components if the drift random effects follow a mixture distribution, default nb.mixt=1.}

\item{mixt.prop}{vector of mixture proportions if the drift random effects follow a mixture distribution, default mixt.prop=1.}

\item{t0}{time origin, default 0.}

\item{X0}{initial value of the process, default X0=0.}

\item{invariant}{1 if the initial value is simulated from the invariant distribution, default 0.01 and X0 is fixed.}

\item{delta}{time step of the simulation (T/N).}

\item{op.plot}{1 if a plot of the trajectories is required, default 0.}

\item{add.plot}{1 for add trajectories to an existing plot}
}
\value{
\item{X}{matrix (M x (N+1)) of the M trajectories. }
\item{phi}{vector (or matrix) of the M simulated random effects.}
}
\description{
Simulation of M independent trajectories of a mixed stochastic differential equation (SDE) with linear drift
 \deqn{dX_j(t)= (\alpha_j- \beta_j X_j(t))dt + \sigma_j a(X_j(t)) dW_j(t)  ,  j=1, ..., M.}
There may be two random effects \eqn{(\alpha_j, \beta_j)} in the drift and one random effect \eqn{\sigma_j} in the diffusion coefficient.
}
\details{
Simulation of M independent trajectories of the SDE (the Brownian motions \eqn{W_j} are independent), with linear drift. There may be one or two
random effects in the drift: 

If drift.random = 0, \eqn{\alpha} and \eqn{\beta} are fixed effects (\eqn{\alpha_j \equiv \alpha} and \eqn{\beta_j \equiv \beta})

If drift.random = 1,  \eqn{\beta} is a fixed effect (\eqn{\beta_j \equiv \beta}), and the drift function is written \eqn{(\alpha_j- \beta X_j(t))}

If drift.random = 2, \eqn{\alpha} is a fixed effect (\eqn{\alpha_j \equiv \alpha}), and the drift function is written \eqn{(\alpha- \beta_j X_j(t))}

If drift.random = c(1,2), both effects are random, and the drift function is written \eqn{(\alpha_j- \beta_j X_j(t))}

Two diffusions are implemented:

Ornstein-Uhlenbeck model (OU): \eqn{a(X_j(t))=1}

Cox-Ingersoll-Ross model (CIR): \eqn{a(X_j(t))=\sqrt{X_j(t)}}

There may be either a fixed or a random effect in the diffusion coefficient:

If diffusion.random = 0, \eqn{\sigma} is a fixed effect (\eqn{\sigma_j \equiv \sigma}). In that case, the random effects in the drift follow a Normal distribution 
or a mixture of Normal distributions.

If diffusion.random = 1, \eqn{\sigma_j} is a random effect. In that case, \eqn{\sigma_j^2} follow an inverse Gamma distribution with parameters diffusion.param=c(shape,scale),
and conditional on \eqn{\sigma_j}, the random effects in the drift follow a Normal distribution with mean mu and variance \eqn{Omega*\sigma_j^2}
}
\references{
This function mixedsde.sim is based on the package sde, function sde.sim. See Simulation and Inference for stochastic differential equation, S.Iacus, \emph{Springer Series in Statistics 2008}
Chapter 2
}
\seealso{
\url{http://cran.r-project.org/package=sde}
}
